Moved logic into a baseclass for MyLitCustomInpector
