+++ Assets/Code/Shaders/Editor/MyCustomShaderInspector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e8e55d7f6676dfd4f9546403c9331e36
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
